+++ tidy3d/plugins/adjoint/components/base.py
@@ -15,6 +15,20 @@ from ....components.base import Tidy3dBaseModel
 from .data.data_array import JaxDataArray, JAX_DATA_ARRAY_TAG
 
 
+# end of the error message when a ``_validate_web_adjoint`` exception is raised
+WEB_ADJOINT_MESSAGE = (
+    "You can still run this simulation through "
+    "'tidy3d.plugins.adjoint.web.run_local' or 'tidy3d.plugins.adjoint.web.run_local' "
+    ", which are similar to 'run' / 'run_async', but "
+    "perform the gradient postprocessing calculation locally after the simulation runs. "
+    "Note that the postprocessing time can become "
+    "quite long (several minutes or more) if these restrictions are exceeded. "
+    "Furthermore, the local versions of 'adjoint' require downloading field data "
+    "inside of the 'input_structures', which can greatly increase the size of data "
+    "needing to be downloaded."
+)
+
+
 class JaxObject(Tidy3dBaseModel):
     """Abstract class that makes a :class:`.Tidy3dBaseModel` jax-compatible through inheritance."""
 
@@ -56,6 +70,10 @@ class JaxObject(Tidy3dBaseModel):
         # TODO: don't use getattr, define this dictionary better
         jax_field_names = self.get_jax_field_names()
         return {key: getattr(self, key) for key in jax_field_names}
+
+    def _validate_web_adjoint(self) -> None:
+        """Run validators for this component, only if using ``tda.web.run()``."""
+        pass
 
     """Methods needed for jax to register arbitrary classes."""