+++ tidy3d/plugins/adjoint/components/geometry.py
@@ -21,7 +21,7 @@ from ....components.monitor import FieldMonitor, PermittivityMonitor
 from ....constants import fp_eps, MICROMETER
 from ....exceptions import AdjointError
 
-from .base import JaxObject
+from .base import JaxObject, WEB_ADJOINT_MESSAGE
 from .types import JaxFloat
 
 # number of integration points per unit wavelength in material
@@ -292,14 +292,17 @@ class JaxPolySlab(JaxGeometry, PolySlab, JaxObject):
             raise AdjointError("'JaxPolySlab' does not support dilation.")
         return val
 
-    @pd.validator("vertices", always=True)
-    def limit_number_of_vertices(cls, val):
+    def _validate_web_adjoint(self) -> None:
+        """Run validators for this component, only if using ``tda.web.run()``."""
+        self._limit_number_of_vertices()
+
+    def _limit_number_of_vertices(self) -> None:
         """Limit the maximum number of vertices."""
-        if len(val) > MAX_NUM_VERTICES:
+        if len(self.vertices_jax) > MAX_NUM_VERTICES:
             raise AdjointError(
-                f"For performance, a maximum of {MAX_NUM_VERTICES} are allowed in 'JaxPolySlab'."
+                f"For performance, a maximum of {MAX_NUM_VERTICES} are allowed in 'JaxPolySlab'. "
+                + WEB_ADJOINT_MESSAGE
             )
-        return val
 
     def edge_contrib(
         self,